refactor: organize files by features
