+++ package.json
@@ -56,6 +56,7 @@
     "inquirer": "^7.0.0",
     "internal-ip": "^6.2.0",
     "jmespath": "^0.15.0",
+    "mongodb": "^4.1.0",
     "ora": "^3.4.0",
     "rimraf": "^3.0.2",
     "shelljs": "^0.8.4",
